[feature] store http query

+++ data/sql/statRequest.sql
@@ -3,5 +3,5 @@ USE rt;
 CREATE TABLE IF NOT EXISTS rt.stat_request(
     `status` smallint(5) NOT NULL,
     `body` varchar(500) NOT NULL DEFAULT '',
-    `duration` smallint(5) NOT NULL
+    `duration` decimal(5,4) NOT NULL
 ) ENGINE InnoDB;
